pyproject.toml
[tool.poetry]
name = "ivy-bot"
version = "0.1.0"
description = "I'm Ivy, a bot created by Ivan to help him automate some things."
authors = ["Ivan Gonzalez <scratchmex@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
PyGithub = "^1.54.1"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
